refactor(agent): separate public authority contract
