+++ packages/agent/src/dkg-agent-cg-resolve.ts
@@ -341,13 +341,13 @@ import {
 import { isTransientBootChainError } from './dkg-agent-boot.js';
 import { createAbortError, runBoundedOperation } from './bounded-operation.js';
 import type {
-  LiveOnChainAccessPolicyState,
   RegisteredContextGraphAuthority,
-} from './internal/context-graph-authority/context-graph-authority.js';
-// Keep the historical dist/dkg-agent-cg-resolve.js type entry point while the
-// authority implementation remains under the package's internal namespace.
+  LiveOnChainAccessPolicyState,
+} from './registered-context-graph-authority.js';
+// Keep the historical dist/dkg-agent-cg-resolve.js type entry point backed by
+// the same stable public contract as the package root.
 export type { RegisteredContextGraphAuthority } from
-  './internal/context-graph-authority/context-graph-authority.js';
+  './registered-context-graph-authority.js';
 import * as diagnostics from './dkg-agent-diagnostics.js';
 import {
   ContextGraphNotFoundError,
