+++ packages/agent/src/internal/context-graph-authority/context-graph-access-policy.ts
@@ -4,7 +4,8 @@ import { createOperationContext, type OperationContext } from
   '@origintrail-official/dkg-core';
 
 import { CHAIN_POLICY_READ_TIMEOUT_MS } from '../../dkg-agent-constants.js';
-import type { LiveOnChainAccessPolicyState } from './context-graph-authority.js';
+import type { LiveOnChainAccessPolicyState } from
+  '../../registered-context-graph-authority.js';
 
 type BoundedPolicyRead<T> =
   | { kind: 'value'; value: T }
