+++ packages/agent/src/registered-context-graph-participant-mutation.ts
@@ -3,7 +3,7 @@
 import type { ChainAdapter, TxResult } from '@origintrail-official/dkg-chain';
 import { ethers } from 'ethers';
 import type { RegisteredContextGraphAuthority } from
-  './internal/context-graph-authority/context-graph-authority.js';
+  './registered-context-graph-authority.js';
 
 export type RegisteredParticipantMutationOperation = 'add' | 'remove';
